changed forbidden exceptions

+++ src/main/java/ipb/pt/safeeat/service/UserService.java
@@ -5,7 +5,7 @@ import ipb.pt.safeeat.model.Cart;
 import ipb.pt.safeeat.model.Restriction;
 import ipb.pt.safeeat.model.User;
 import ipb.pt.safeeat.repository.*;
-import ipb.pt.safeeat.utility.NotAllowedConstants;
+import ipb.pt.safeeat.utility.ForbiddenConstants;
 import ipb.pt.safeeat.utility.NotFoundConstants;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -44,8 +44,8 @@ public class UserService {
 
         User current = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
-        if(!current.isAdmin() && !current.equals(user))
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN, NotAllowedConstants.FORBIDDEN_USER);
+        if (!current.isAdmin() && !current.equals(user))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, ForbiddenConstants.FORBIDDEN_USER);
 
         return user;
     }
@@ -77,6 +77,7 @@ public class UserService {
                         () -> new ResponseStatusException(HttpStatus.NOT_FOUND, NotFoundConstants.RESTRICTION_NOT_FOUND)));
             }
         }
+
         return restrictions;
     }
 
@@ -97,7 +98,7 @@ public class UserService {
         User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
         if (!user.getId().equals(userDto.getId()))
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Cannot update other user");
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, ForbiddenConstants.FORBIDDEN_USER);
 
         User byEmail = userRepository.findByEmail(userDto.getEmail()).orElse(null);
 
@@ -117,12 +118,18 @@ public class UserService {
         User user = userRepository.findById(id).orElseThrow(
                 () -> new ResponseStatusException(HttpStatus.NOT_FOUND, NotFoundConstants.USER_NOT_FOUND));
 
+        User current = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!current.isAdmin() && !current.equals(user))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, ForbiddenConstants.FORBIDDEN_USER);
+
         if (!user.getRestaurants().isEmpty())
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Cannot delete user with restaurants");
 
         paymentRepository.deleteAll(user.getPayments());
         addressRepository.deleteAll(user.getAddresses());
         cartRepository.delete(user.getCart());
+
         userRepository.deleteById(id);
     }
 }
